Package uploads: Got CLI working with API
